refactor: moved types and interface from service to utils
