--- src/service/constans.ts
@@ -1,18 +1,0 @@
-import { TdisciplinesButtons } from './utils/types';
-
-export const disciplines: TdisciplinesButtons[] =
-  [
-    { name: 'Skateboarding', to: '/skateboarding' },
-    { name: 'Street art', to: '/street art' },
-    { name: 'Parkour', to: '/parkour' },
-    { name: 'Workout', to: '/workout' },
-    { name: 'Freerun', to: '/freerun' },
-    { name: 'Hip-Hop dance', to: '/hiphop' },
-    { name: 'Trieking', to: '/trieking' },
-    { name: 'Rap', to: '/rap' },
-    { name: 'Breaking', to: '/breaking' },
-    { name: 'BMX', to: '/bmx' },
-    { name: 'Scoot', to: '/scoot' },
-  ];
-
-export const ADD_USER = 'ADD_USER';
