+++ src/service/hooks/hooks.ts
@@ -1,5 +1,5 @@
 import { useDispatch, useSelector } from 'react-redux';
-import { AppDispatch, RootState } from '../types';
+import { AppDispatch, RootState } from '../../utils/types/storeTypes';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 export const useAppSelector = useSelector.withTypes<RootState>();
