--- src/service/types.ts
@@ -1,9 +1,0 @@
-import { rootReducer } from './slice/root';
-import { setupStore } from './store';
-
-export type RootState = ReturnType<typeof rootReducer>;
-export type AppStore = ReturnType<typeof setupStore>;
-export type AppDispatch = AppStore['dispatch'];
-
-
-
