--- src/service/utils/cookes.ts
@@ -1,43 +1,0 @@
-export function getCookie(name: string) {
-  const matches = document.cookie.match(
-    new RegExp(
-      '(?:^|; )' +
-      name.replace(/([.$?*|{}()[]\\\/\+^])/g, '\\$1') +
-      '=([^;]*)',
-    ),
-  );
-  return matches ? decodeURIComponent(matches[1]) : undefined;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function setCookie(name: string, value: string | number | boolean, props?: any) {
-  props = props || {};
-
-  let exp = props.expires;
-  if (typeof exp == 'number' && exp) {
-    const d = new Date();
-    d.setTime(d.getTime() + exp * 1000);
-    exp = props.expires = d;
-  }
-
-  if (exp && exp instanceof Date && exp.toUTCString) {
-    props.expires = exp.toUTCString();
-  }
-
-  value = encodeURIComponent(value);
-  let updatedCookie = name + '=' + value;
-
-  for (const propName in props) {
-    updatedCookie += '; ' + propName;
-    const propValue = props[propName];
-
-    if (propValue !== true) {
-      updatedCookie += '=' + propValue;
-    }
-  }
-  document.cookie = updatedCookie;
-}
-
-export function deleteCookie(name: string) {
-  setCookie(name, '', { expires: -1 });
-}
