+++ src/utils/interface/disciplineInterface.ts
@@ -1,0 +1,19 @@
+export interface IDiscipline {
+  name: string,
+  imagesUrl: string[],
+  description: string,
+  rules: string,
+  isfull?: boolean,
+}
+
+export interface IDisciplineState {
+  disciplines: IDiscipline[],
+  currentDiscipline: IDiscipline,
+  isLoading: boolean,
+  error: string | undefined,
+}
+
+export interface IDisciplinesButtons {
+  name: string,
+  to: string,
+}
